Add files via upload
This C++ code defines a 'Student' class with two member variables: 'name' and 'grade'. The program prompts the user to enter information for 5 students. The user must enter a name that is not empty, does not contain spaces, and starts with a capital letter. If the user does not provide valid input for a student, the program informs the user and prompts them to enter the information again.

Once the program has valid input for 5 students, it determines which student has the highest grade. It accomplishes this by looping through the 'Student' objects and comparing their grade variable to a 'highestGrade' variable that is initially set as the first student. If a 'Student' object has a higher grade than the current 'highestGrade', the 'highestGrade' variable is updated to the new value, and the name variable of the corresponding Student object is saved in a 'highestName' variable.

Finally, the program prints the name and grade of the student with the highest grade.